requirements.txt
asgiref==3.5.0
certifi==2021.10.8
chardet==3.0.4
Django==4.0.3
django-autocomplete-light==3.9.1
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
django-easy-select2==1.5.8
django-extensions==3.1.5
django-filter==21.1
django-grappelli==3.0.3
djangorestframework==3.13.1
idna==2.7
Pillow==6.2.0
psycopg2==2.9.3
pytz==2021.3
requests==2.20.1
six==1.11.0
sqlparse==0.4.2
urllib3==1.24.3
opencv-python-headless==4.5.5.62